# 🚀 Optimize CI Pipeline: 3.3x Faster Tests with 100% Success Rate