--- a/.Renviron
+++ b/.Renviron
@@ -1,0 +1,1 @@
+GIS_DIR=/work/gis