--- a/DESCRIPTION
+++ b/DESCRIPTION
@@ -36,6 +36,7 @@ Imports:
     logger,
     matrixStats,
     rlang,
+    sf,
     pbapply
 Depends:
     tidyverse,