--- a/Dockerfile.ci
+++ b/Dockerfile.ci
@@ -97,11 +97,12 @@ RUN if [ "$DOWNLOAD_GIS_DATA" = "true" ]; then \
                 "gs://crea-data/gis/population/gpw_v4_population_density_adjusted_to_2015_unwpp_country_totals_rev11_2020_30_sec.tif" \
                 $GIS_DIR/population/ && \
             echo "Downloading specific boundary data..." && \
+            mkdir -p $GIS_DIR/boundaries/gadm && \
             gsutil -o "Credentials:gs_service_key_file=/tmp/gcs-key.json" -m cp \
                 "gs://crea-data/gis/boundaries/gadm/gadm410_0_low_rgeos.RDS" \
                 "gs://crea-data/gis/boundaries/gadm/gadm410_1_low_rgeos.RDS" \
                 "gs://crea-data/gis/boundaries/gadm/gadm410_2_low_rgeos.RDS" \
-                $GIS_DIR/boundaries/ && \
+                $GIS_DIR/boundaries/gadm/ && \
             echo "=== GCS DOWNLOAD COMPLETE ===" && \
             ls -la $GIS_DIR/population/ && \
             ls -la $GIS_DIR/boundaries/ && \
@@ -125,6 +126,7 @@ RUN apt-get update && apt-get install -y \
 RUN Rscript -e 'pak::pkg_install(c("doParallel"), ask = FALSE)'
 
 # Install R package dependencies and the package itself
+# .Renviron is excluded via .dockerignore to avoid overriding GIS_DIR
 COPY . /tmp/creahia
 WORKDIR /tmp/creahia
 RUN Rscript -e 'pak::local_install_deps(".", ask = FALSE)' && \